pyproject.toml
[tool.poetry]
name = "crossbar_llm"
version = "0.1.0"
description = ""
authors = ["Melih DARCAN <57872471+MelihDarcanxyz@users.noreply.github.com>"]
license = "License :: OSI Approved :: GNU General Public License v3 (GPLv3)"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"

langchain = "^0.2.1"
langchain-anthropic = "^0.1.13"
langchain-community = "^0.2.1"
langchain-google-genai = "^1.0.5"
langchain-groq = "^0.1.4"
neo4j = "^5.20.0"
numpy = "^1.26.4"
openai = "^1.30.2"
pandas = "^2.2.2"
pydantic = "^2.7.1"
python-dotenv = "^1.0.1"
langchain-nvidia-ai-endpoints = "^0.2.0"
wrapt-timeout-decorator = "^1.5.1"
langchain-openai = "^0.1.20"
fastapi = "^0.115.5"

[tool.poetry.dev-dependencies]

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true
without = ["dev"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
